split out printer of results
